Move Fragment to avoid recursive dependencies; Update react to latest for examples; Remove shrinkwrap
