+++ package.json
@@ -73,9 +73,9 @@
     "prettier": "^1.5.2",
     "pretty-bytes": "^4.0.2",
     "raw-loader": "^0.5.1",
-    "react": "^16.0.0",
+    "react": "^16.5.0",
     "react-ace": "^5.4.0",
-    "react-dom": "^16.0.0",
+    "react-dom": "^16.5.0",
     "react-router": "^4.2.0",
     "react-router-dom": "^4.2.2",
     "source-map-loader": "^0.1.6",
