+++ src/util/Fragment.js
@@ -1,0 +1,3 @@
+import React from 'react';
+
+export const Fragment = React.Fragment || 'div';
